EGL: Corrected colour keying for non-RGB32 images.
The colour key should be loaded in RGB888, and not the source picture format;
the pixel shader doesn't know the source format. Use dfb_pixel_to_color().

DFB and GL are free to differ in the bottom-most bits when expanding a colour
channel from 6 or 5 bits to 8 bits. The bottom 2 or 3 bits must be ignored
when comparing a pixel to the colour-key. GLSL 1.0 doesn't have bitmasks, so
this is done with integer arithmetic.