SaWMan: Return an error if looked-up window is NULL
DirectFB CoreWindow stores a reference to SaWManWindow in the
window_data member. If the CoreWindow is in destroyed state then the
SaWManWindow does not exist anymore and the pointer is set to a NULL
pointer.

This commit ensures that SaWManWindow_Lookup() returns an error if the
window is in a destroyed state and therefore no related SaWManWindow
exist.